+++ src/service.rs
@@ -29,6 +29,7 @@ impl Service {
             LlmBackendSettings::Zhipu { model, .. } => model.clone(),
             LlmBackendSettings::Volcengine { model, .. } => model.clone(),
             LlmBackendSettings::Tencent { model, .. } => model.clone(),
+            LlmBackendSettings::Longcat { model, .. } => model.clone(),
         };
 
         Ok(Self { client, model })